\name{la2rcpp-package}
\alias{la2rcpp-package}
\alias{la2rcpp}
\docType{package}
\title{\packageTitle{la2rcpp}}
\description{\packageDescription{la2rcpp}}
\section{Package Content}{\packageIndices{la2rcpp}}
\author{\packageAuthor{la2rcpp}}
\section{Maintainer}{\packageMaintainer{la2rcpp}}
\keyword{package}
